Refactor cols in hook
