+++ context/app/static/js/components/detailPage/CollectionDatasetsTable/hooks.js
@@ -1,17 +1,35 @@
 import { useSearchHits } from 'js/hooks/useSearchData';
 import { getIDsQuery } from 'js/helpers/queries';
+import {
+  lastModifiedTimestampCol,
+  organCol,
+  dataTypesCol,
+  statusCol,
+} from 'js/components/detailPage/derivedEntities/columns';
 
-function useCollectionsDatasets({ ids, sourceFields }) {
+const columns = [
+  organCol,
+  dataTypesCol,
+  lastModifiedTimestampCol,
+  {
+    id: 'created_by_user_displayname',
+    label: 'Contact',
+    renderColumnCell: ({ created_by_user_displayname }) => created_by_user_displayname,
+  },
+  statusCol,
+];
+
+function useCollectionsDatasets({ ids }) {
   const query = {
     query: {
       ...getIDsQuery(ids),
     },
-    _source: sourceFields,
+    _source: ['hubmap_id', 'entity_type', ...columns.map((c) => c.id)],
     size: 10000,
   };
 
   const { searchHits: datasets } = useSearchHits(query);
-  return datasets;
+  return { columns, datasets };
 }
 
 export { useCollectionsDatasets };
